Add ability to run .coffee files
Other than the "works on my system" setup, not sure the best way to go
about testing this.  Compiling it and running it on my system works, all
vows continue to work.

In addition, I've also added typos to the `require('coffee-script')`
line to force it to act like CoffeeScript isn't available.  It continues
to run the `*.js` files and ignores the `*.coffee` files.