Add ability to circumvent `addBatch`
This makes it possible to skip the calls to `addBatch` by using optional
parameters inside the `describe` method.  This is a small change that
allows the developer to focus on the code of the tests, thinking of it
think of them in terms of data structure instead of Vows' API.

The less that the Vows API is present in a test case, the less
distracting it is.  With this, you could choose to focus solely on the
structure of the test and ignore Vows is present by importing only
`describe`.

This is an entirely backwards compatible change that does not effect any
other code.  It is fully tested inside the existing tests.