Attempt to call exit(1) when some specs fail
The 'vows' executable was returning an exit code of 0 even when some
tests were broken or raising errors.  This problem was caused by the
script exiting without ever issuing a 'drain' event on 'process.stdout',
and thus not running the drain handler that called 'process.exit'.

To fix this bug, I first needed to find out whether it was possible to
detect distinguish fully-drained stdout from undrained stdout.  It turns
out that calling "process.stdout.write('')" will return false if I/O
is pending.  I established this using the following CoffeeScript code:

  # Very approximate "repeat string about N times" function.
  repeat = (s,n) -> if n <= 1 then s else tmp=repeat(s,n/2); tmp+tmp

  # Send a big output request followed by a 0-length one.
  r1=process.stdout.write(repeat("a", 200000)); r2=process.stdout.write('')

  # Both I/O calls report that output is still pending.
  coffee> r1
  false
  coffee> r2
  false

This isn't a very nice or elegant solution, and there's probably a nicer
way to do this with Node.js.