[test v0.6] Make `assert-test.js` v0.6-compatible
Since joyent/node@8c8d518, `assert` module is a function. When you look
at lib/vows/suite.js:126, it checks if topic is a function and then, if
it is, it executes it and passes results down to vows. Of course, it
fails hard in this case, since `assert` module is a function (resulting
in error without a stack trace).