[v0.6 fix] Properly inspect errors
This is related and blocked by cloudhead/eyes.js#12.

Gist of what's going on: `Error` properties are non-enumerable in `node
v0.6`, so `Object.keys` which is being used in `eyes.js` won't list them
(we need to use `Object.getOwnPropertyNames`). This results in errors
being outputted as `{}`, for example in `assert.isNull(err)` statement.