Fix bare assert() to work properly
According to MDN, arrow functions don't bind their own `this` or
`arguments` (among other things) - that would possibly explain why
this failed, but not why it totally worked when I stepped through it
in a debugger. So in conclusion I honestly have zero idea why this
works, but the tests didn't pass before, and now they do. So ship it I
guess?

Fixes #23